qcacld-3.0: sap: Fix invalid casting from a void to an enumerator.
This fixes the following warnings:

../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1552:26: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                bss_complete->status = (eSapStatus) context;
                                       ^~~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1571:4: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                        (eSapStatus) context;
                        ^~~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1575:4: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                        (eSapStatus)context);
                        ^~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1580:30: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                if (eSAP_STATUS_SUCCESS == (eSapStatus)context) {
                                           ^~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1588:37: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                } else if (eSAP_STATUS_FAILURE == (eSapStatus)context) {
                                                  ^~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1596:4: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                        (eSapStatus) context;
                        ^~~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1671:30: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                reassoc_complete->status = (eSapStatus) context;
                                           ^~~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1715:27: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                disassoc_comp->status = (eSapStatus) context;
                                        ^~~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1733:26: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                key_complete->status = (eSapStatus) context;
                                       ^~~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1783:4: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                        (eSapStatus) context;
                        ^~~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1789:4: warning: cast to smaller integer type 'eSapStatus' from 'void *' [-Wvoid-pointer-to-enum-cast]
                        (eSapStatus) context;
                        ^~~~~~~~~~~~~~~~~~~~
../drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c:1866:46: warning: cast to smaller integer type 'QDF_STATUS' from 'void *' [-Wvoid-pointer-to-enum-cast]
                sap_ap_event.sapevt.ch_change_rsp_status = (QDF_STATUS)context;
                                                           ^~~~~~~~~~~~~~~~~~~

[saurabhchardereal: Apply same patch to qca-wifi-host-cmn/hif/src/hif_napi.c]

+++ drivers/staging/qcacld-3.0/core/sap/src/sap_fsm.c
@@ -2952,7 +2952,7 @@ QDF_STATUS sap_signal_hdd_event(ptSapContext sap_ctx,
 		sap_ap_event.sapHddEventCode = eSAP_START_BSS_EVENT;
 		bss_complete = &sap_ap_event.sapevt.sapStartBssCompleteEvent;
 
-		bss_complete->status = (eSapStatus) context;
+		bss_complete->status = (eSapStatus)(long) context;
 		bss_complete->staId = sap_ctx->sap_sta_id;
 
 		QDF_TRACE(QDF_MODULE_ID_SAP, QDF_TRACE_LEVEL_INFO_HIGH,
@@ -2972,16 +2972,16 @@ QDF_STATUS sap_signal_hdd_event(ptSapContext sap_ctx,
 	case eSAP_DFS_NO_AVAILABLE_CHANNEL:
 		sap_ap_event.sapHddEventCode = sap_hddevent;
 		sap_ap_event.sapevt.sapStopBssCompleteEvent.status =
-			(eSapStatus) context;
+			(eSapStatus)(long)context;
 		break;
 	case eSAP_ACS_SCAN_SUCCESS_EVENT:
 		sap_handle_acs_scan_event(sap_ctx, &sap_ap_event,
-			(eSapStatus)context);
+			(eSapStatus)(long)context);
 		break;
 	case eSAP_ACS_CHANNEL_SELECTED:
 		sap_ap_event.sapHddEventCode = sap_hddevent;
 		acs_selected = &sap_ap_event.sapevt.sap_ch_selected;
-		if (eSAP_STATUS_SUCCESS == (eSapStatus)context) {
+		if (eSAP_STATUS_SUCCESS == (eSapStatus)(long)context) {
 			acs_selected->pri_ch = sap_ctx->acs_cfg->pri_ch;
 			acs_selected->ht_sec_ch = sap_ctx->acs_cfg->ht_sec_ch;
 			acs_selected->ch_width = sap_ctx->acs_cfg->ch_width;
@@ -2989,7 +2989,7 @@ QDF_STATUS sap_signal_hdd_event(ptSapContext sap_ctx,
 				sap_ctx->acs_cfg->vht_seg0_center_ch;
 			acs_selected->vht_seg1_center_ch =
 				sap_ctx->acs_cfg->vht_seg1_center_ch;
-		} else if (eSAP_STATUS_FAILURE == (eSapStatus)context) {
+		} else if (eSAP_STATUS_FAILURE == (eSapStatus)(long)context) {
 			acs_selected->pri_ch = 0;
 		}
 		break;
@@ -2997,7 +2997,7 @@ QDF_STATUS sap_signal_hdd_event(ptSapContext sap_ctx,
 	case eSAP_STOP_BSS_EVENT:
 		sap_ap_event.sapHddEventCode = eSAP_STOP_BSS_EVENT;
 		sap_ap_event.sapevt.sapStopBssCompleteEvent.status =
-			(eSapStatus) context;
+			(eSapStatus)(long)context;
 		break;
 
 	case eSAP_STA_ASSOC_EVENT:
@@ -3067,7 +3067,7 @@ QDF_STATUS sap_signal_hdd_event(ptSapContext sap_ctx,
 		chaninfo->rate_flags = csr_roaminfo->chan_info.rate_flags;
 
 		reassoc_complete->wmmEnabled = csr_roaminfo->wmmEnabledSta;
-		reassoc_complete->status = (eSapStatus) context;
+		reassoc_complete->status = (eSapStatus)(long)context;
 		reassoc_complete->timingMeasCap = csr_roaminfo->timingMeasCap;
 		reassoc_complete->ecsa_capable = csr_roaminfo->ecsa_capable;
 		reassoc_complete->ampdu = csr_roaminfo->ampdu;
@@ -3110,7 +3110,7 @@ QDF_STATUS sap_signal_hdd_event(ptSapContext sap_ctx,
 			disassoc_comp->reason = eSAP_MAC_INITATED_DISASSOC;
 
 		disassoc_comp->statusCode = csr_roaminfo->statusCode;
-		disassoc_comp->status = (eSapStatus) context;
+		disassoc_comp->status = (eSapStatus)(long)context;
 		disassoc_comp->rssi = csr_roaminfo->rssi;
 		disassoc_comp->rx_rate = csr_roaminfo->rx_rate;
 		disassoc_comp->tx_rate = csr_roaminfo->tx_rate;
@@ -3128,7 +3128,7 @@ QDF_STATUS sap_signal_hdd_event(ptSapContext sap_ctx,
 		sap_ap_event.sapHddEventCode = eSAP_STA_SET_KEY_EVENT;
 		key_complete =
 			&sap_ap_event.sapevt.sapStationSetKeyCompleteEvent;
-		key_complete->status = (eSapStatus) context;
+		key_complete->status = (eSapStatus)(long)context;
 		qdf_copy_macaddr(&key_complete->peerMacAddr,
 				 &csr_roaminfo->peerMac);
 		break;
@@ -3187,13 +3187,13 @@ QDF_STATUS sap_signal_hdd_event(ptSapContext sap_ctx,
 	case eSAP_DISCONNECT_ALL_P2P_CLIENT:
 		sap_ap_event.sapHddEventCode = eSAP_DISCONNECT_ALL_P2P_CLIENT;
 		sap_ap_event.sapevt.sapActionCnf.actionSendSuccess =
-			(eSapStatus) context;
+			(eSapStatus)(long)context;
 		break;
 
 	case eSAP_MAC_TRIG_STOP_BSS_EVENT:
 		sap_ap_event.sapHddEventCode = eSAP_MAC_TRIG_STOP_BSS_EVENT;
 		sap_ap_event.sapevt.sapActionCnf.actionSendSuccess =
-			(eSapStatus) context;
+			(eSapStatus)(long)context;
 		break;
 
 	case eSAP_UNKNOWN_STA_JOIN:
